# Some more Q4_K and Q5_K speedup on CUDA