Quite minor on the modern card (RTX-4080), but fairly significant on the older card that I have available (GTX-1660). For `Q5_K` the same bit fiddling for computing the scales is added as in #2322. For both, we gain somewhat if we do 2 dot products per 32 quants in the dot product kernel. Both changes apply to `LLAMA_CUDA_FORCE_DMMV=OFF`.

* TG-128 in ms/t on on GTX-1660 (6GB VRAM, so no 13B results, for `Q5_K_S` 32 layers fit on the GPU).

| Quants  | Master | PR | Speedup |
|--:|--:|--:|--:|
| Q4_K_S | 29.5 | 25.6 | 15.2% |
| Q5_K_S | 42.8 | 35.5 |  20.6% |  

* TG-128 in ms/t on RTX-4080:

| Quants  | Master | PR | Speedup |
|--:|--:|--:|--:| 
| Q4_K_S - 7B | 8.40 | 8.23 | 2.1% |
| Q5_K_S - 7B | 10.42 | 9.52 |  9.5% |
| Q4_K_S - 13B |  14.64 | 14.38 | 1.8% |
| Q5_K_S - 13B |  17.89 | 16.80 | 6.5% |