@JohannesGaessler 2023-07-23
@ikawrakow which k-quants do you plan to touch in the immediate future (next 2-3 days)? I want to finalize my matrix matrix multiplication PR and as part of that PR I'm decoupling the actual dot products from the code that loads the primitives from the blocks. It would be convenient for me to know your plans in order to minimize rebasing.