@ikawrakow 2023-07-23
> @ikawrakow which k-quants do you plan to touch in the immediate future (next 2-3 days)? I want to finalize my matrix matrix multiplication PR and as part of that PR I'm decoupling the actual dot products from the code that loads the primitives from the blocks. It would be convenient for me to know your plans in order to minimize rebasing.

I have another PR basically ready. It fixes the `Q4_K` and `Q5_K` dot products for a block size of 64. Then I should be done for now with the CUDA code.